added events to `Session`
A session will now emit certain events during its lifecycle.

It will emit `ledger-entry` whenever a new ledger entry is added, with the
details of the entry.

It will also emit `agent-start`, `agent-stopping`, `agent-step` whenever
any agent is started/stopped or takes a step,
passing the agent ID and its current state. It will also emit
`agent-stop` once an agent's response is finalized, with the agent ID
and the response.

Finally, it will emit `heartbeat` event every time an
agent takes a step, or the agent's `heartbeat` method is manually
called. This is used on the platform to monitor agent activity and to
timeout inactive sessions.