added support for `tool_choice` option
Agents can specify `tool_choice` option to control how the model
responds. With `auto`, it chooses by itself whether to make a tool call
and which tool to call. With `none`, it won't call any tools. With
`required`, it will always make a tool call instead of a normal
response. It's also possible to specify a concrete tool that the model
must call.